+++ AS-pratica/app/Http/Controllers/LivroController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Livro;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class LivroController extends Controller
 {
@@ -15,6 +16,8 @@ class LivroController extends Controller
 
     public function create()
     {
+        Gate::authorize('create', Livro::class);
+
         return view('livros.create');
     }
 
@@ -35,6 +38,8 @@ class LivroController extends Controller
 
     public function edit($id)
     {
+        Gate::authorize('update', Livro::class);
+
         $livro = Livro::findOrFail($id);
         return view('livros.edit', compact('livro'));
     }
@@ -54,6 +59,8 @@ class LivroController extends Controller
 
     public function destroy($id)
     {
+        Gate::authorize('delete', Livro::class);
+
         $livro = Livro::findOrFail($id);
         $livro->delete();
         return redirect('livros')->with('success', 'Livro deleted successfully.');
